Inject DevicePowerManager
