+++ Client-Raspi/src/main/org/kabieror/elwasys/raspiclient/devices/IDevicePowerMeasurementHandler.java
@@ -1,0 +1,5 @@
+package org.kabieror.elwasys.raspiclient.devices;
+
+public interface IDevicePowerMeasurementHandler {
+    void onPowerMeasurementAvailable(int deviceId, double currentPowerConsumption);
+}
